refactor(re-member): 도서관 페이지에 input 적용
