+++ apps/re-member/src/components/library/index.ts
@@ -1,4 +1,4 @@
-export { default as LibrarySearchBar } from "./LibrarySearchBar";
+export { default as LibrarySearchInput } from "./LibrarySearchInput";
 export { default as LibraryFilter } from "./LibraryFilter";
 export { default as LibraryBookList } from "./LibraryBookList";
 export { default as LibraryBookItem } from "./LibraryBookItem";
